+++ src/core/request/engines/xhr.ts
@@ -11,7 +11,7 @@ import XMLHttpRequest from 'xhr2';
 //#endif
 
 import { IS_NODE } from 'core/env';
-import Then from 'core/then';
+import AbortablePromise from 'core/promise/abortable';
 import { isOnline } from 'core/net';
 
 import Response from 'core/request/response';
@@ -87,9 +87,9 @@ const request: RequestEngine = (params) => {
 		xhr.setRequestHeader('Content-Type', contentType);
 	}
 
-	return new Then<Response>(async (resolve, reject, onAbort) => {
+	return new AbortablePromise<Response>(async (resolve, reject, onAbort) => {
 		const
-			{status} = await Then.resolve(isOnline(), p.parent);
+			{status} = await AbortablePromise.resolve(isOnline(), p.parent);
 
 		if (!status) {
 			return reject(new RequestError('offline', {